requirements.txt
tweepy
selenium
python-dotenv